Fix most* of eric's stupidity *plexiGlass subclass is unfinished and we need a default(empty) constructor for shelterAnimal
