+++ plexiGlass.java
@@ -1,0 +1,6 @@
+public class plexiGlass extends shelterAnimal{
+    private boolean plexiGlass;
+
+    public plexiGlass
+    
+}
